refactor: move authn-related code to authn package
