+++ build-logic/src/main/kotlin/conventions/kotlin.gradle.kts
@@ -52,12 +52,15 @@ kotlin {
 
 detekt {
     parallel = true
+    autoCorrect = true
     buildUponDefaultConfig = true
 
     config.setFrom(File(rootProject.projectDir, ".config/detekt/detekt.yaml"))
 }
 
 tasks.withType<Detekt>().configureEach {
+    exclude("**/resources/**", "**/build/**", "**/generated/**")
+
     reports {
         sarif.required = true
     }
