+++ server/src/main/kotlin/org/elaastix/server/authn/AuthenticationHolder.kt
@@ -17,29 +17,21 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package org.elaastix.server.tmp
+package org.elaastix.server.authn
 
-import jakarta.annotation.PostConstruct
-import org.elaastix.server.users.UserRepository
 import org.elaastix.server.users.entities.User
-import org.springframework.context.annotation.Profile
-import org.springframework.data.domain.Pageable
+import org.springframework.http.HttpStatus
+import org.springframework.security.core.context.SecurityContextHolder
 import org.springframework.stereotype.Component
+import org.springframework.web.client.HttpClientErrorException
 
 @Component
-@Profile("develop")
-class InitUsers(
-    val userRepository: UserRepository
-) {
-    @PostConstruct
-    fun doInitUsers() {
-        if (userRepository.count() == 0L) {
-            userRepository.save(User())
-            userRepository.save(User())
-            userRepository.save(User())
-        }
+class AuthenticationHolder {
+    val isAuthenticated: Boolean
+        get() = SecurityContextHolder.getContext().authentication?.isAuthenticated == true
 
-        @Suppress("MagicNumber")
-        userRepository.findAll(Pageable.ofSize(3)).forEach { println("Test user 1: ${it.id}") }
-    }
+    val authenticatedUser: User?
+        get() = SecurityContextHolder.getContext().authentication?.principal as? User
+
+    fun User?.required(): User = this ?: throw HttpClientErrorException(HttpStatus.UNAUTHORIZED)
 }
